+++ docs/examples/basic_context/implicit_context.py
@@ -1,4 +1,4 @@
-# Copyright 2020, OpenTelemetry Authors
+# Copyright The OpenTelemetry Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,4 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__version__ = "0.15.dev0"
+from opentelemetry import baggage, trace
+from opentelemetry.sdk.trace import TracerProvider
+
+trace.set_tracer_provider(TracerProvider())
+tracer = trace.get_tracer(__name__)
+
+with tracer.start_span(name="root span") as root_span:
+    ctx = baggage.set_baggage("foo", "bar")
+
+print("Global context baggage: {}".format(baggage.get_all()))
+print("Span context baggage: {}".format(baggage.get_all(context=ctx)))