+++ instrumentation/opentelemetry-instrumentation-tornado/src/opentelemetry/instrumentation/tornado/__init__.py
@@ -50,6 +50,7 @@ from opentelemetry import configuration, context, propagators, trace
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.instrumentation.tornado.version import __version__
 from opentelemetry.instrumentation.utils import (
+    extract_attributes_from_object,
     http_status_to_canonical_code,
     unwrap,
 )
@@ -71,7 +72,17 @@ def get_excluded_urls():
     return ExcludeList(urls)
 
 
+def get_traced_request_attrs():
+    attrs = configuration.Configuration().TORNADO_TRACED_REQUEST_ATTRS or ""
+    if attrs:
+        attrs = [attr.strip() for attr in attrs.split(",")]
+    else:
+        attrs = []
+    return attrs
+
+
 _excluded_urls = get_excluded_urls()
+_traced_attrs = get_traced_request_attrs()
 
 
 class TornadoInstrumentor(BaseInstrumentor):
@@ -196,7 +207,7 @@ def _get_attributes_from_request(request):
     if request.remote_ip:
         attrs["net.peer.ip"] = request.remote_ip
 
-    return attrs
+    return extract_attributes_from_object(request, _traced_attrs, attrs)
 
 
 def _get_operation_name(handler, request):
@@ -211,12 +222,16 @@ def _start_span(tracer, handler, start_time) -> _TraceContext:
             _get_header_from_request_headers, handler.request.headers,
         )
     )
+
     span = tracer.start_span(
         _get_operation_name(handler, handler.request),
         kind=trace.SpanKind.SERVER,
-        attributes=_get_attributes_from_request(handler.request),
         start_time=start_time,
     )
+    if span.is_recording():
+        attributes = _get_attributes_from_request(handler.request)
+        for key, value in attributes.items():
+            span.set_attribute(key, value)
 
     activation = tracer.use_span(span, end_on_exit=True)
     activation.__enter__()
@@ -248,15 +263,16 @@ def _finish_span(tracer, handler, error=None):
     if not ctx:
         return
 
-    if reason:
-        ctx.span.set_attribute("http.status_text", reason)
-    ctx.span.set_attribute("http.status_code", status_code)
-    ctx.span.set_status(
-        Status(
-            canonical_code=http_status_to_canonical_code(status_code),
-            description=reason,
+    if ctx.span.is_recording():
+        if reason:
+            ctx.span.set_attribute("http.status_text", reason)
+        ctx.span.set_attribute("http.status_code", status_code)
+        ctx.span.set_status(
+            Status(
+                canonical_code=http_status_to_canonical_code(status_code),
+                description=reason,
+            )
         )
-    )
 
     ctx.activation.__exit__(*finish_args)
     context.detach(ctx.token)