+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/export/in_memory_metrics_exporter.py
@@ -15,7 +15,11 @@
 import threading
 from typing import Sequence
 
-from . import MetricRecord, MetricsExporter, MetricsExportResult
+from opentelemetry.sdk.metrics.export import (
+    MetricRecord,
+    MetricsExporter,
+    MetricsExportResult,
+)
 
 
 class InMemoryMetricsExporter(MetricsExporter):