+++ opentelemetry-sdk/tests/trace/export/test_export.py
@@ -21,6 +21,8 @@ from logging import WARNING
 from unittest import mock
 
 from opentelemetry import trace as trace_api
+from opentelemetry.configuration import Configuration
+from opentelemetry.context import Context
 from opentelemetry.sdk import trace
 from opentelemetry.sdk.trace import export
 
@@ -100,6 +102,23 @@ class TestSimpleExportSpanProcessor(unittest.TestCase):
 
         self.assertListEqual(["xxx", "bar", "foo"], spans_names_list)
 
+    def test_on_start_accepts_context(self):
+        # pylint: disable=no-self-use
+        tracer_provider = trace.TracerProvider()
+        tracer = tracer_provider.get_tracer(__name__)
+
+        exporter = MySpanExporter([])
+        span_processor = mock.Mock(
+            wraps=export.SimpleExportSpanProcessor(exporter)
+        )
+        tracer_provider.add_span_processor(span_processor)
+
+        context = Context()
+        span = tracer.start_span("foo", context=context)
+        span_processor.on_start.assert_called_once_with(
+            span, parent_context=context
+        )
+
     def test_simple_span_processor_not_sampled(self):
         tracer_provider = trace.TracerProvider(
             sampler=trace.sampling.ALWAYS_OFF
@@ -136,6 +155,11 @@ def _create_start_and_end_span(name, span_processor):
 
 
 class TestBatchExportSpanProcessor(unittest.TestCase):
+    def tearDown(self) -> None:
+        # reset global state of configuration object
+        # pylint: disable=protected-access
+        Configuration._reset()
+
     @mock.patch.dict(
         "os.environ",
         {
@@ -155,6 +179,23 @@ class TestBatchExportSpanProcessor(unittest.TestCase):
         self.assertEqual(batch_span_processor.schedule_delay_millis, 2)
         self.assertEqual(batch_span_processor.max_export_batch_size, 3)
         self.assertEqual(batch_span_processor.export_timeout_millis, 4)
+
+    def test_on_start_accepts_parent_context(self):
+        # pylint: disable=no-self-use
+        my_exporter = MySpanExporter(destination=[])
+        span_processor = mock.Mock(
+            wraps=export.BatchExportSpanProcessor(my_exporter)
+        )
+        tracer_provider = trace.TracerProvider()
+        tracer_provider.add_span_processor(span_processor)
+        tracer = tracer_provider.get_tracer(__name__)
+
+        context = Context()
+        span = tracer.start_span("foo", context=context)
+
+        span_processor.on_start.assert_called_once_with(
+            span, parent_context=context
+        )
 
     def test_shutdown(self):
         spans_names_list = []